apps/worker/requirements.txt
temporalio==1.8.0
httpx==0.27.2
psycopg[binary]==3.2.13
psycopg-pool==3.2.3
langgraph==0.2.39
langgraph-checkpoint-postgres==2.0.25
pydantic==2.9.2
pydantic-settings==2.5.2
prometheus-client==0.21.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-httpx==0.48b0
setuptools==79.0.1
